Move files around and clean up complex methods
