+++ src/Provider/GitHub/GitHubPRService.ts
@@ -7,7 +7,7 @@ import { TIME_ZONE, USER_AGENT } from '../../Config/constants/defaults';
 import { Log } from '../../Logger';
 import { Diff } from '../../Git/@types/PatchTypes';
 import { API_PAGE_SIZE_LIMIT } from '../constants/github.provider.constant';
-import { getPatch } from '../utils/patchProcessor';
+import { getPatch } from '../../Git/utils/patchProcessor';
 import { CommitStatus } from './CommitStatus';
 import { IGitHubPRService } from './IGitHubPRService';
 import {
